Avoid ambiguous overload. Cleanup support functions.
