+++ src/main/scala/com/emotioncity/soriento/loadbyname/AnyRichODatabaseDocumentImpl.scala
