+++ src/main/scala/com/emotioncity/soriento/loadbyname/ClassToNameFunctions.scala
