+++ src/test/scala/com/emotioncity/soriento/loadbyname/ODBUtil.scala
@@ -1,10 +1,10 @@
 package com.emotioncity.soriento.loadbyname
 
+import com.orientechnologies.orient.client.remote.OServerAdmin
 import com.orientechnologies.orient.core.db.document.ODatabaseDocumentTx
 import com.orientechnologies.orient.core.metadata.schema.OClass
 import com.orientechnologies.orient.core.record.impl.ODocument
 import com.orientechnologies.orient.core.sql.query.OSQLSynchQuery
-
 import scala.collection.JavaConverters._
 
 /**
@@ -64,5 +64,70 @@ object ODBUtil{
       }
     }
   }
+
+
+  def ensureNewRemoteDB(host: String,
+                        dbName: String,
+                        user: String = "admin",
+                        pass: String = "admin",
+                        remoteDBStorage: String = "plocal",
+                        dbType: String = "document",
+                        dropIfExists: Boolean = false): Unit = {
+    val server = new OServerAdmin(s"remote:${host}/${dbName}").connect(user, pass)
+
+    try {
+      val exists = server.listDatabases.asScala.contains(dbName)
+      if (exists) {
+        if (dropIfExists) {
+          println("DROP!")
+          server.dropDatabase(dbName)
+          server.createDatabase(dbName, dbType, remoteDBStorage)
+        }
+        // Already exists
+      }
+      else {
+        server.createDatabase(dbName, dbType, remoteDBStorage)
+      }
+    }
+    finally {
+      server.close()
+    }
+  }
+
+
+  def makeTestDB() = {
+    if (true) {
+
+      val db: ODatabaseDocumentTx = new ODatabaseDocumentTx("memory:testdb")
+      db.create()
+      db
+
+    } else {
+      // Remote DB
+      // This is useful for checking that serialization actually happens.
+      val dbHost = "orientdb"
+      val dbName = "stu"
+      val dbUser = "root"
+      val dbPass = ""
+
+      ensureNewRemoteDB(dbHost, dbName, dbUser, dbPass, remoteDBStorage = "plocal", dropIfExists = true)
+      new ODatabaseDocumentTx(s"remote:${dbHost}/${dbName}").open(dbUser, dbPass)
+    }
+  }
+
+  def withDropDB(dbFactory: => ODatabaseDocumentTx)(func: ODatabaseDocumentTx => Unit): Unit = {
+    val db = dbFactory
+    try {
+      func(db)
+    } finally {
+      if (!db.isClosed) {
+        if (!db.isActiveOnCurrentThread) {
+          db.activateOnCurrentThread()
+        }
+        db.drop()
+      }
+    }
+  }
+
 
 }
